Change endpoint url to match the provided format
Change endpoint URLs from api/v1/allRides/ to api/v1/rides as provided
Will add an api endpoint to request to join a ride
Fix import errors
